Move common visualizer types to visualizer_types.h
