+++ src/visualizer_common.cc
@@ -6,6 +6,7 @@
  */
  
 #include "visualizer.h"
+#include "visualizer_types.h"
 #include "visualizer_common.h"
 #include "visualizer_circuit.h"
 #include "visualizer_interaction.h"
